@@ -14,13 +14,20 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.1.1",
     "@apollo/react-testing": "^3.1.3",
+    "@babel/plugin-proposal-private-property-in-object": "^7.16.0",
     "@babel/preset-react": "^7.0.0",
     "@date-io/date-fns": "1.x",
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.4.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
     "@material-ui/pickers": "^3.2.10",
     "@material-ui/styles": "^4.9.0",
+    "@mui/icons-material": "^5.1.0",
+    "@mui/lab": "^5.0.0-alpha.54",
+    "@mui/material": "^5.1.0",
+    "@mui/styles": "^5.1.0",
     "@rails/actioncable": "^6.0.0-alpha",
     "@rails/activestorage": "^6.0.0-alpha",
     "@rails/ujs": "^6.0.0-alpha",
@@ -39,7 +46,6 @@
     "dom-to-image": "^2.6.0",
     "dompurify": "^2.0.12",
     "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.15.1",
     "eslint-plugin-react-hooks": "4.0.0",
     "file-loader": "^6.2.0",
     "graphql": "^14.5.6",
@@ -55,11 +61,11 @@
     "material-ui-confirm": "^2.1.1",
     "moment-timezone": "^0.5.33",
     "prop-types": "^15.7.2",
-    "react": "^16.9.0",
+    "react": "17.0.0",
     "react-apollo": "^3.1.1",
     "react-awesome-query-builder": "^3.0.0",
     "react-csv": "^2.0.3",
-    "react-dom": "^16.9.0",
+    "react-dom": "17.0.0",
     "react-easy-crop": "^3.3.1",
     "react-email-editor": "^1.2.0",
     "react-ga": "^2.7.0",
@@ -84,15 +90,16 @@
   "devDependencies": {
     "@babel/core": "^7.12.3",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
-    "@storybook/addon-actions": "6.0.28",
-    "@storybook/addon-essentials": "6.0.28",
-    "@storybook/addon-links": "6.0.28",
-    "@storybook/react": "6.0.28",
+    "@storybook/addon-actions": "6.3.0",
+    "@storybook/addon-essentials": "6.3.0",
+    "@storybook/addon-links": "6.3.0",
+    "@storybook/react": "6.3.0",
     "@testing-library/dom": "^8.2.0",
     "@testing-library/jest-dom": "^5.5.0",
     "@testing-library/react": "^10.0.3",
     "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^13.2.1",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
     "babel-eslint": "^10.0.3",
     "babel-jest": "^27.2.4",
     "babel-loader": "^8.2.1",
