@@ -1,43 +1,43 @@
 /* eslint-disable */
 module.exports = function(api) {
-  var validEnv = ["development", "test", "production"];
+  var validEnv = ['development', 'test', 'production'];
   var currentEnv = api.env();
-  var isDevelopmentEnv = api.env("development");
-  var isProductionEnv = api.env("production");
-  var isTestEnv = api.env("test");
+  var isDevelopmentEnv = api.env('development');
+  var isProductionEnv = api.env('production');
+  var isTestEnv = api.env('test');
 
   if (!validEnv.includes(currentEnv)) {
     throw new Error(
-      "Please specify a valid `NODE_ENV` or " +
+      'Please specify a valid `NODE_ENV` or ' +
         '`BABEL_ENV` environment variables. Valid values are "development", ' +
         '"test", and "production". Instead, received: ' +
         JSON.stringify(currentEnv) +
-        "."
+        '.'
     );
   }
 
   return {
     presets: [
       isTestEnv && [
-        require("@babel/preset-env").default,
+        require('@babel/preset-env').default,
         {
           targets: {
-            node: "current"
+            node: 'current'
           }
         }
       ],
       (isProductionEnv || isDevelopmentEnv) && [
-        require("@babel/preset-env").default,
+        require('@babel/preset-env').default,
         {
           forceAllTransforms: true,
-          useBuiltIns: "entry",
+          useBuiltIns: 'entry',
           corejs: 3,
           modules: false,
-          exclude: ["transform-typeof-symbol"]
+          exclude: ['transform-typeof-symbol']
         }
       ],
       [
-        require("@babel/preset-react").default,
+        require('@babel/preset-react').default,
         {
           development: isDevelopmentEnv || isTestEnv,
           useBuiltIns: true
@@ -45,25 +45,27 @@ module.exports = function(api) {
       ]
     ].filter(Boolean),
     plugins: [
-      require("babel-plugin-macros"),
-      require("@babel/plugin-syntax-dynamic-import").default,
-      isTestEnv && require("babel-plugin-dynamic-import-node"),
+      require('babel-plugin-macros'),
+      require('@babel/plugin-syntax-dynamic-import').default,
+      isTestEnv && require('babel-plugin-dynamic-import-node'),
       isDevelopmentEnv && require('react-refresh/babel'),
-      require("@babel/plugin-transform-destructuring").default,
+      require('@babel/plugin-transform-destructuring').default,
       [
-        require("@babel/plugin-proposal-class-properties").default,
+        require('@babel/plugin-proposal-class-properties').default,
         {
           loose: true
         }
       ],
+      [require('@babel/plugin-proposal-private-methods').default, { loose: true }],
+      [require('@babel/plugin-proposal-private-property-in-object').default, { loose: true }],
       [
-        require("@babel/plugin-proposal-object-rest-spread").default,
+        require('@babel/plugin-proposal-object-rest-spread').default,
         {
           useBuiltIns: true
         }
       ],
       [
-        require("@babel/plugin-transform-runtime").default,
+        require('@babel/plugin-transform-runtime').default,
         {
           helpers: false,
           regenerator: true,
@@ -71,13 +73,13 @@ module.exports = function(api) {
         }
       ],
       [
-        require("@babel/plugin-transform-regenerator").default,
+        require('@babel/plugin-transform-regenerator').default,
         {
           async: false
         }
       ],
       isProductionEnv && [
-        require("babel-plugin-transform-react-remove-prop-types").default,
+        require('babel-plugin-transform-react-remove-prop-types').default,
         {
           removeImport: true
         }
